# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GeoKrety Stats API
  version: 3.1.0
  description: "Read-only API for GeoKrety statistics,\n    resource exploration.\n\
    \nThe `/api/v3/...` contract is JSON REST based: every JSON response uses\ntop-level\
    \ `data`, `meta`, and `links`, while resource objects expose `id`,\n`type`, `attributes`,\
    \ optional `relationships`, and resource-scoped\n`links.self`.\nCollection endpoints\
    \ are cursor-first for now: clients send `limit`, follow\n`links.next`, and continue\
    \ while `meta.has_more` is true. Search endpoints\nrequire a `q` query parameter\
    \ of at least 2 characters.\nThe public API surface was intentionally reduced\
    \ during the big-bang\nmigration: redundant activity, stats, alias, and visualization\
    \ endpoints were\nremoved so the stable contract centers on core resources and\
    \ scoped\ncollections.\n\n## GeoKret Public Identifiers (GKID)\n\nGeoKrety endpoints\
    \ accept public identifiers in three forms:\n- **Numeric decimal**: `1`, `123`,\
    \ `65535` (direct GKID number)\n- **Bare hexadecimal**: `0001`, `00FF`, `FFFF`\
    \ (hex without the `GK` prefix)\n- **GK-prefixed hexadecimal**: `GK0001`, `GK00FF`,\
    \ `GKFFFF`\n  (padded hex representation)\n\nParsing is case-insensitive. Digit-only\
    \ inputs without a leading zero are\ntreated as decimal, while zero-padded digit-only\
    \ inputs are treated as\nhexadecimal.\n\nAll forms resolve to the same GeoKret.\
...
servers:
...
paths:
...
  /api/v3/geokrety:
    get:
      tags:
      - Geokrety
      operationId: listGeokrety
      summary: List GeoKrety
      parameters:
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/CursorParam'
      - $ref: '#/components/parameters/GeokretIdsParam'
      responses:
        '200':
          description: Cursor-paginated GeoKrety collection.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/JsonRestCollectionDocument'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/GeokretResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v3/geokrety/search:
    get:
      tags:
      - Geokrety
      operationId: searchGeokrety
      summary: Search GeoKrety
      parameters:
      - $ref: '#/components/parameters/SearchQueryParam'
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: Matching GeoKrety.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/JsonRestCollectionDocument'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/GeokretResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v3/countries:
    get:
      tags:
      - Countries
      operationId: listCountries
      summary: List countries
      parameters:
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/CursorParam'
      - $ref: '#/components/parameters/CountryIdsParam'
      responses:
        '200':
          description: Cursor-paginated country collection.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/JsonRestCollectionDocument'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/CountryResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v3/waypoints/search:
    get:
      tags:
      - Waypoints
      operationId: searchWaypoints
      summary: Search waypoints
      parameters:
      - $ref: '#/components/parameters/SearchQueryParam'
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: Matching waypoint collection.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/JsonRestCollectionDocument'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/WaypointResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v3/waypoints/{code}:
    get:
      tags:
      - Waypoints
      operationId: getWaypoint
      summary: Waypoint details
      parameters:
      - $ref: '#/components/parameters/WaypointCodeParam'
      responses:
        '200':
          description: Single waypoint resource document.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/JsonRestSingleDocument'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/WaypointResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v3/users:
    get:
      tags:
      - Users
      operationId: listUsers
      summary: List users
      parameters:
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/CursorParam'
      - $ref: '#/components/parameters/UserIdsParam'
      responses:
        '200':
          description: Cursor-paginated user collection.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/JsonRestCollectionDocument'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/UserResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v3/users/search:
    get:
      tags:
      - Users
      operationId: searchUsers
      summary: Search users
      parameters:
      - $ref: '#/components/parameters/SearchQueryParam'
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: Matching user collection.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/JsonRestCollectionDocument'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/UserResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v3/pictures:
    get:
      tags:
      - Pictures
      operationId: listPictures
      summary: List pictures
      parameters:
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/CursorParam'
      - $ref: '#/components/parameters/PictureIdsParam'
      responses:
        '200':
          description: Cursor-paginated picture collection.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/JsonRestCollectionDocument'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/PictureResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    LimitParam:
      in: query
      name: limit
      description: Maximum number of items to return. Cursor-paginated endpoints default
        to 20 and reject values above the endpoint-specific maximum.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 20
    CursorParam:
      in: query
      name: cursor
      description: Opaque cursor token returned in `links.next`. Do not construct
        or modify this value on the client.
      required: false
      schema:
        type: string
...
    SearchQueryParam:
      in: query
      name: q
      description: Search text. Must be at least 2 characters long.
      required: true
      schema:
        type: string
        minLength: 2
      example: gk
    GeokretIdsParam:
      in: query
      name: ids
      description: Optional comma-separated list of public GeoKret identifiers. When
        provided, the endpoint returns exactly the requested resources and ignores
        cursor pagination.
      required: false
      schema:
        type: string
        pattern: ^(?:\s*(?:[Gg][Kk])?0*[1-9A-Fa-f][0-9A-Fa-f]*\s*)(?:,\s*(?:[Gg][Kk])?0*[1-9A-Fa-f][0-9A-Fa-f]*\s*)*$
      example: GK0001,GK00FF
...
    CountryIdsParam:
      in: query
      name: ids
      description: Optional comma-separated list of ISO-3166 alpha-2 country codes.
        When provided, the endpoint returns exactly the requested country resources
        and ignores cursor pagination.
      required: false
      schema:
        type: string
        pattern: ^\s*[A-Za-z]{2}(?:\s*,\s*[A-Za-z]{2})*\s*$
      example: PL,DE,FR
    UserIdsParam:
      in: query
      name: ids
      description: Optional comma-separated list of numeric user identifiers. When
        provided, the endpoint returns exactly the requested user resources and ignores
        cursor pagination.
      required: false
      schema:
        type: string
        pattern: ^\s*[1-9][0-9]*(?:\s*,\s*[1-9][0-9]*)*\s*$
      example: 1,42,99
    PictureIdsParam:
      in: query
      name: ids
      description: Optional comma-separated list of numeric picture identifiers. When
        provided, the endpoint returns exactly the requested picture resources and
        ignores cursor pagination.
      required: false
      schema:
        type: string
        pattern: ^\s*[1-9][0-9]*(?:\s*,\s*[1-9][0-9]*)*\s*$
      example: 5,6,7
...
    WaypointCodeParam:
      in: path
      name: code
      description: Waypoint code or source-specific cache identifier.
      required: true
      schema:
        type: string
        minLength: 2
        maxLength: 32
        pattern: ^[A-Za-z0-9_-]{2,32}$
      example: GC12345
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JsonRestErrorDocument'
          example:
            error:
              code: BAD_REQUEST
              message: invalid identifier
            timestamp: '2026-03-17T09:12:34Z'
        application/xml:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: invalid identifier
            timestamp: '2026-03-17T09:12:34Z'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JsonRestErrorDocument'
          example:
            error:
              code: NOT_FOUND
              message: resource not found
            timestamp: '2026-03-17T09:12:34Z'
        application/xml:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: resource not found
            timestamp: '2026-03-17T09:12:34Z'
    InternalServerError:
      description: The server failed to read or prepare the requested data.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JsonRestErrorDocument'
          example:
            error:
              code: INTERNAL_ERROR
              message: failed to fetch resource
            timestamp: '2026-03-17T09:12:34Z'
        application/xml:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: failed to fetch resource
            timestamp: '2026-03-17T09:12:34Z'
...
  schemas:
    JsonRestResourceLinks:
      type: object
      additionalProperties: false
      properties:
        self:
          type: string
    JsonRestDocumentLinks:
      type: object
      additionalProperties: false
      properties:
        self:
          type: string
        next:
          type: string
    JsonRestIdentifier:
      type: object
      required:
      - type
      - id
      properties:
        type:
          type: string
        id:
          type: string
    JsonRestRelationship:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/JsonRestIdentifier'
        links:
          $ref: '#/components/schemas/JsonRestResourceLinks'
    JsonRestMeta:
      type: object
      required:
      - execution_time_ms
      properties:
        execution_time_ms:
          type: integer
          description: Milliseconds spent handling the request.
        limit:
          type: integer
          description: Requested collection size for cursor collections.
        has_more:
          type: boolean
          description: Indicates whether `links.next` can be followed.
        page:
          type: integer
        per_page:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
    JsonRestErrorPayload:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
        message:
          type: string
    JsonRestErrorDocument:
      type: object
      required:
      - error
      - timestamp
      properties:
        error:
          $ref: '#/components/schemas/JsonRestErrorPayload'
        timestamp:
          type: string
          format: date-time
    JsonRestResource:
      type: object
      required:
      - id
      - type
      properties:
        id:
          type: string
        type:
          type: string
        attributes:
          type: object
          additionalProperties: true
        relationships:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/JsonRestRelationship'
        links:
          $ref: '#/components/schemas/JsonRestResourceLinks'
    JsonRestSingleDocument:
      type: object
      required:
      - data
      - meta
      - links
      properties:
        data:
          $ref: '#/components/schemas/JsonRestResource'
        meta:
          $ref: '#/components/schemas/JsonRestMeta'
        links:
          $ref: '#/components/schemas/JsonRestDocumentLinks'
    JsonRestCollectionDocument:
      type: object
      required:
      - data
      - meta
      - links
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/JsonRestResource'
        meta:
          $ref: '#/components/schemas/JsonRestMeta'
        links:
          $ref: '#/components/schemas/JsonRestDocumentLinks'
    GeokretResource:
      allOf:
      - $ref: '#/components/schemas/JsonRestResource'
      - type: object
        properties:
          type:
            type: string
            enum:
            - geokret
...
    UserResource:
      allOf:
      - $ref: '#/components/schemas/JsonRestResource'
      - type: object
        properties:
          type:
            type: string
            enum:
            - user
    PictureResource:
      allOf:
      - $ref: '#/components/schemas/JsonRestResource'
      - type: object
        properties:
          type:
            type: string
            enum:
            - picture
    CountryResource:
      allOf:
      - $ref: '#/components/schemas/JsonRestResource'
      - type: object
        properties:
          type:
            type: string
            enum:
            - country
    WaypointResource:
      allOf:
      - $ref: '#/components/schemas/JsonRestResource'
      - type: object
        properties:
          type:
            type: string
            enum:
            - waypoint
...
    ErrorResponse:
      type: object
      xml:
        name: ErrorResponse
      required:
      - error
      - timestamp
      properties:
        error:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
...